Tested with different values for getting the max. gas cost for transaction types
This is important to see how many transactions could be added in a single block for a contract's method. E.g. in Load contract, only 1 tx(heavy) can be added with 'setArray' method with count value: 2650 as it is hitting the block gas limit. So, while running the CLI for heavy txs, one needs to know that only this value of count when sending so many tx in one go. So, then not more than 1 tx would be added even when running in batches (100 each, say).